input to output path and analyisis control

+++ p2/KafkaProducer/src/main/java/com/bigdata/luka/kafkaproducer/model/KafkaEvent.java
@@ -1,13 +1,11 @@
 package com.bigdata.luka.kafkaproducer.model;
 
-import java.math.BigDecimal;
-
 public interface KafkaEvent {
 
     String getKey();
 
     String getVehicleId();
 
-    BigDecimal getTimestamp();
+    Long getTimestamp();
 
 }
